Moved jdbc url injection to postgresql-ds.xml instead of standalone.xml. Moved Arquillian extension to util package
