+++ qa/integration-tests-engine/src/test/java/org/operaton/bpm/integrationtest/util/ArquillianEventObserver.java
@@ -1,4 +1,4 @@
-package org.operaton.bpm.integrationtest.deployment.cfg;
+package org.operaton.bpm.integrationtest.util;
 
 import org.jboss.arquillian.container.spi.ContainerRegistry;
 import org.jboss.arquillian.core.api.annotation.Observes;
