+++ qa/integration-tests-engine/src/test/resources/META-INF/services/org.jboss.arquillian.core.spi.LoadableExtension
@@ -1,1 +1,1 @@
-org.operaton.bpm.integrationtest.deployment.cfg.ArquillianLifecycleObserverExtension
+org.operaton.bpm.integrationtest.util.ArquillianLifecycleObserverExtension
